split placer into modules
