+++ deps/ui/src/views/basic/circle_view.rs
@@ -27,7 +27,7 @@ impl CircleView {
     pub fn set_radius(&mut self, radius: impl ToF32) -> &mut Self {
         let radius = radius.to_f32();
 
-        if radius == self.radius {
+        if (radius - self.radius).abs() < f32::EPSILON {
             return self;
         }
